Plugin structure change
- App Engine Plugin (Decider, can apply)
- App Engine Core Plugin
- Standard Plugin
- Flexible Plugin

This change also changes how extension are created, using
nested extension instead of evaluation closures.
